pyproject.toml
[tool.poetry]
name = "onidataapi"
version = "0.1.0"
description = ""
authors = ["alfmorais <alfredneto@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
markdown = "^3.5.2"
django-filter = "^23.5"
python-dateutil = "^2.8.2"


[tool.poetry.group.test.dependencies]
ipdb = "^0.13.13"
pre-commit = "^3.6.0"
pytest = "^8.0.0"
pytest-django = "^4.8.0"
freezegun = "^1.4.0"
factory-boy = "^3.3.0"
pytest-factoryboy = "^2.6.0"
pytest-vcr = "^1.0.2"

[tool.pytest.ini_options]
pythonpath = ". src"
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["tests_*.py", "*_test.py", "testing/python/*.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
